+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Serposcope changelog
 
+## UPCOMING
+
+* fix class COM not found bug (affected: Windows + PHP > 5.4.5)
+* detect uninstalled COM extension on install (affected: Windows + PHP > 5.4.5)
+* fix undefined PHP_MAJOR_VERSION / PHP_MINOR_VERSION (affected: PHP < 5.2.7)
+* fix last check not displayed on main page
+* more verbose error message on import fail
+* fix go ninja go install link
+
 ## 1.0.1 - 28/03/2013
 
 * proxies bulk import (contributed by @512banque http://twitter.com/512banque)